+++ src/main/java/webserver/constant/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.constant;
 
 public enum HttpMethod {
