+++ src/main/java/webserver/handler/HttpRequestHandler.java
@@ -1,4 +1,7 @@
-package webserver;
+package webserver.handler;
+
+import webserver.request.HttpRequest;
+import webserver.response.HttpResponse;
 
 public interface HttpRequestHandler {
